app/requirements.txt
flask==2.0.1
werkzeug==2.0.3
boto3==1.38.1
python-dotenv==0.19.0
gunicorn==20.1.0 